--- test/kube2e/upgrade/upgrade_test.go
@@ -1,426 +1,0 @@
-package upgrade_test
-
-import (
-	"bytes"
-	"context"
-	"fmt"
-	"io/ioutil"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
-	"text/template"
-	"time"
-
-	"github.com/solo-io/gloo/test/kube2e/upgrade"
-
-	exec_utils "github.com/solo-io/go-utils/testutils/exec"
-	"github.com/solo-io/k8s-utils/kubeutils"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
-
-	"github.com/solo-io/skv2/codegen/util"
-
-	"github.com/solo-io/gloo/projects/gloo/cli/pkg/helpers"
-	"github.com/solo-io/solo-kit/pkg/api/v1/clients"
-
-	"github.com/ghodss/yaml"
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-	"github.com/solo-io/gloo/projects/gloo/cli/pkg/cmd/version"
-	"github.com/solo-io/gloo/projects/gloo/pkg/defaults"
-	"github.com/solo-io/gloo/test/kube2e"
-	"github.com/solo-io/go-utils/versionutils"
-	"github.com/solo-io/k8s-utils/testutils/helper"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime"
-	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
-)
-
-const namespace = defaults.GlooSystem
-
-var _ = Describe("Kube2e: Upgrade Tests", func() {
-
-	var (
-		crdDir     string
-		chartUri   string
-		ctx        context.Context
-		cancel     context.CancelFunc
-		testHelper *helper.SoloTestHelper
-
-		// whether to set validation webhook's failurePolicy=Fail
-		strictValidation bool
-
-		// Versions to upgrade from
-		// ex: current branch is 1.13.10 - this would be the latest patch release of 1.12
-		LastPatchMostRecentMinorVersion *versionutils.Version
-
-		// ex:current branch is 1.13.10 - this would be 1.13.9
-		CurrentPatchMostRecentMinorVersion *versionutils.Version
-	)
-
-	// setup for all tests
-	BeforeEach(func() {
-		ctx, cancel = context.WithCancel(context.Background())
-
-		cwd, err := os.Getwd()
-		Expect(err).NotTo(HaveOccurred())
-		testHelper, err = helper.NewSoloTestHelper(func(defaults helper.TestConfig) helper.TestConfig {
-			defaults.RootDir = filepath.Join(cwd, "../../..")
-			defaults.HelmChartName = "gloo"
-			defaults.InstallNamespace = namespace
-			defaults.Verbose = true
-			return defaults
-		})
-		Expect(err).NotTo(HaveOccurred())
-
-		crdDir = filepath.Join(util.GetModuleRoot(), "install", "helm", "gloo", "crds")
-		chartUri = filepath.Join(testHelper.RootDir, testHelper.TestAssetDir, testHelper.HelmChartName+"-"+testHelper.ChartVersion()+".tgz")
-		strictValidation = false
-
-		LastPatchMostRecentMinorVersion, CurrentPatchMostRecentMinorVersion, err = upgrade.GetUpgradeVersions(ctx, "gloo")
-		Expect(err).NotTo(HaveOccurred())
-	})
-
-	Describe("Upgrading from a previous gloo version to current version", func() {
-		Context("When upgrading from LastPatchMostRecentMinorVersion to PR version of gloo", func() {
-			BeforeEach(func() {
-				installGloo(testHelper, LastPatchMostRecentMinorVersion.String(), strictValidation)
-			})
-			AfterEach(func() {
-				uninstallGloo(testHelper, ctx, cancel)
-			})
-			//It("Used for local testing to check base case upgrades", func() {
-			//	baseUpgradeTest(ctx, crdDir, LastPatchMostRecentMinorVersion.String(), testHelper, chartUri, strictValidation)
-			//})
-			It("uses helm to update validationServerGrpcMaxSizeBytes without errors", func() {
-				updateSettingsWithoutErrors(ctx, crdDir, testHelper, chartUri, strictValidation)
-			})
-			It("uses helm to add a second gateway-proxy in a separate namespace without errors", func() {
-				addSecondGatewayProxySeparateNamespaceTest(crdDir, testHelper, chartUri, strictValidation)
-			})
-		})
-		Context("When upgrading from CurrentPatchMostRecentMinorVersion to PR version of gloo", func() {
-			BeforeEach(func() {
-				installGloo(testHelper, CurrentPatchMostRecentMinorVersion.String(), strictValidation)
-			})
-			AfterEach(func() {
-				uninstallGloo(testHelper, ctx, cancel)
-			})
-			It("uses helm to update validationServerGrpcMaxSizeBytes without errors", func() {
-				updateSettingsWithoutErrors(ctx, crdDir, testHelper, chartUri, strictValidation)
-			})
-			It("uses helm to add a second gateway-proxy in a separate namespace without errors", func() {
-				addSecondGatewayProxySeparateNamespaceTest(crdDir, testHelper, chartUri, strictValidation)
-			})
-		})
-	})
-
-	Context("Validation webhook upgrade tests", func() {
-		var cfg *rest.Config
-		var err error
-		var kubeClientset kubernetes.Interface
-
-		BeforeEach(func() {
-			cfg, err = kubeutils.GetConfig("", "")
-			Expect(err).NotTo(HaveOccurred())
-			kubeClientset, err = kubernetes.NewForConfig(cfg)
-			Expect(err).NotTo(HaveOccurred())
-			strictValidation = true
-		})
-
-		Context("When upgrading from LastPatchMostRecentMinorVersion to PR version of gloo", func() {
-			BeforeEach(func() {
-				installGloo(testHelper, LastPatchMostRecentMinorVersion.String(), strictValidation)
-			})
-			AfterEach(func() {
-				uninstallGloo(testHelper, ctx, cancel)
-			})
-			It("sets validation webhook caBundle on install and upgrade", func() {
-				updateValidationWebhookTests(ctx, crdDir, kubeClientset, testHelper, chartUri, false)
-			})
-		})
-
-		Context("When upgrading from CurrentPatchMostRecentMinorVersion to PR version of gloo", func() {
-			BeforeEach(func() {
-				installGloo(testHelper, CurrentPatchMostRecentMinorVersion.String(), strictValidation)
-			})
-			AfterEach(func() {
-				uninstallGloo(testHelper, ctx, cancel)
-			})
-			It("sets validation webhook caBundle on install and upgrade", func() {
-				updateValidationWebhookTests(ctx, crdDir, kubeClientset, testHelper, chartUri, false)
-			})
-		})
-	})
-})
-
-// ===================================
-// Repeated Test Code
-// ===================================
-// Based case test for local runs to help narrow down failures
-func baseUpgradeTest(ctx context.Context, crdDir string, startingVersion string, testHelper *helper.SoloTestHelper, chartUri string, strictValidation bool) {
-	By(fmt.Sprintf("should start with gloo version %s", startingVersion))
-	Expect(fmt.Sprintf("v%s", getGlooServerVersion(ctx, testHelper.InstallNamespace))).To(Equal(startingVersion))
-
-	// upgrade to the gloo version being tested
-	upgradeGloo(testHelper, chartUri, crdDir, strictValidation, nil)
-
-	By("should have upgraded to the gloo version being tested")
-	Expect(getGlooServerVersion(ctx, testHelper.InstallNamespace)).To(Equal(testHelper.ChartVersion()))
-}
-
-func updateSettingsWithoutErrors(ctx context.Context, crdDir string, testHelper *helper.SoloTestHelper, chartUri string, strictValidation bool) {
-	By("should start with the settings.invalidConfigPolicy.invalidRouteResponseCode=404")
-	client := helpers.MustSettingsClient(ctx)
-	settings, err := client.Read(testHelper.InstallNamespace, defaults.SettingsName, clients.ReadOpts{})
-	Expect(err).To(BeNil())
-	Expect(settings.GetGloo().GetInvalidConfigPolicy().GetInvalidRouteResponseCode()).To(Equal(uint32(404)))
-
-	upgradeGloo(testHelper, chartUri, crdDir, strictValidation, []string{
-		"--set", "settings.replaceInvalidRoutes=true",
-		"--set", "settings.invalidConfigPolicy.invalidRouteResponseCode=400",
-		"--set", "gateway.validation.validationServerGrpcMaxSizeBytes=5000000",
-	})
-
-	By("should have updated to settings.invalidConfigPolicy.invalidRouteResponseCode=400")
-	settings, err = client.Read(testHelper.InstallNamespace, defaults.SettingsName, clients.ReadOpts{})
-	Expect(err).To(BeNil())
-	Expect(settings.GetGloo().GetInvalidConfigPolicy().GetInvalidRouteResponseCode()).To(Equal(uint32(400)))
-	Expect(settings.GetGateway().GetValidation().GetValidationServerGrpcMaxSizeBytes().GetValue()).To(Equal(int32(5000000)))
-}
-
-func addSecondGatewayProxySeparateNamespaceTest(crdDir string, testHelper *helper.SoloTestHelper, chartUri string, strictValidation bool) {
-	const externalNamespace = "other-ns"
-	requiredSettings := map[string]string{
-		"gatewayProxies.proxyExternal.disabled":              "false",
-		"gatewayProxies.proxyExternal.namespace":             externalNamespace,
-		"gatewayProxies.proxyExternal.service.type":          "NodePort",
-		"gatewayProxies.proxyExternal.service.httpPort":      "31500",
-		"gatewayProxies.proxyExternal.service.httpsPort":     "32500",
-		"gatewayProxies.proxyExternal.service.httpNodePort":  "31500",
-		"gatewayProxies.proxyExternal.service.httpsNodePort": "32500",
-	}
-
-	var settings []string
-	for key, val := range requiredSettings {
-		settings = append(settings, "--set")
-		settings = append(settings, strings.Join([]string{key, val}, "="))
-	}
-
-	runAndCleanCommand("kubectl", "create", "ns", externalNamespace)
-	defer runAndCleanCommand("kubectl", "delete", "ns", externalNamespace)
-
-	upgradeGloo(testHelper, chartUri, crdDir, strictValidation, settings)
-
-	// Ensures deployment is created for both default namespace and external one
-	// Note - name of external deployments is kebab-case of gatewayProxies NAME helm value
-	Eventually(func() (string, error) {
-		return exec_utils.RunCommandOutput(testHelper.RootDir, false,
-			"kubectl", "get", "deployment", "-A")
-	}, "10s", "1s").Should(
-		And(ContainSubstring("gateway-proxy"),
-			ContainSubstring("proxy-external")))
-
-	// Ensures service account is created for the external namespace
-	Eventually(func() (string, error) {
-		return exec_utils.RunCommandOutput(testHelper.RootDir, false,
-			"kubectl", "get", "serviceaccount", "-n", externalNamespace)
-	}, "10s", "1s").Should(ContainSubstring("gateway-proxy"))
-}
-
-func updateValidationWebhookTests(ctx context.Context, crdDir string, kubeClientset kubernetes.Interface, testHelper *helper.SoloTestHelper, chartUri string, strictValidation bool) {
-	webhookConfigClient := kubeClientset.AdmissionregistrationV1().ValidatingWebhookConfigurations()
-	secretClient := kubeClientset.CoreV1().Secrets(testHelper.InstallNamespace)
-
-	By("the webhook caBundle should be the same as the secret's root ca value")
-	webhookConfig, err := webhookConfigClient.Get(ctx, "gloo-gateway-validation-webhook-"+testHelper.InstallNamespace, metav1.GetOptions{})
-	Expect(err).NotTo(HaveOccurred())
-	secret, err := secretClient.Get(ctx, "gateway-validation-certs", metav1.GetOptions{})
-	Expect(err).NotTo(HaveOccurred())
-	Expect(webhookConfig.Webhooks[0].ClientConfig.CABundle).To(Equal(secret.Data[corev1.ServiceAccountRootCAKey]))
-
-	upgradeGloo(testHelper, chartUri, crdDir, strictValidation, nil)
-
-	By("the webhook caBundle and secret's root ca value should still match after upgrade")
-	webhookConfig, err = webhookConfigClient.Get(ctx, "gloo-gateway-validation-webhook-"+testHelper.InstallNamespace, metav1.GetOptions{})
-	Expect(err).NotTo(HaveOccurred())
-	secret, err = secretClient.Get(ctx, "gateway-validation-certs", metav1.GetOptions{})
-	Expect(err).NotTo(HaveOccurred())
-	Expect(webhookConfig.Webhooks[0].ClientConfig.CABundle).To(Equal(secret.Data[corev1.ServiceAccountRootCAKey]))
-}
-
-// ===================================
-// Util methods
-// ===================================
-func getGlooServerVersion(ctx context.Context, namespace string) (v string) {
-	glooVersion, err := version.GetClientServerVersions(ctx, version.NewKube(namespace, ""))
-	Expect(err).To(BeNil())
-	Expect(len(glooVersion.GetServer())).To(Equal(1))
-	for _, container := range glooVersion.GetServer()[0].GetKubernetes().GetContainers() {
-		if v == "" {
-			v = container.OssTag
-		} else {
-			Expect(container.OssTag).To(Equal(v))
-		}
-	}
-	return v
-}
-
-func makeUnstructured(yam string) *unstructured.Unstructured {
-	jsn, err := yaml.YAMLToJSON([]byte(yam))
-	ExpectWithOffset(1, err).NotTo(HaveOccurred())
-	runtimeObj, err := runtime.Decode(unstructured.UnstructuredJSONScheme, jsn)
-	ExpectWithOffset(1, err).NotTo(HaveOccurred())
-	return runtimeObj.(*unstructured.Unstructured)
-}
-
-func makeUnstructuredFromTemplateFile(fixtureName string, values interface{}) *unstructured.Unstructured {
-	tmpl, err := template.ParseFiles(fixtureName)
-	ExpectWithOffset(1, err).NotTo(HaveOccurred())
-	var b bytes.Buffer
-	err = tmpl.Execute(&b, values)
-	ExpectWithOffset(1, err).NotTo(HaveOccurred())
-	return makeUnstructured(b.String())
-}
-
-func installGloo(testHelper *helper.SoloTestHelper, fromRelease string, strictValidation bool) {
-	valueOverrideFile, cleanupFunc := kube2e.GetHelmValuesOverrideFile()
-	defer cleanupFunc()
-
-	// construct helm args
-	var args = []string{"install", testHelper.HelmChartName}
-
-	runAndCleanCommand("helm", "repo", "add", testHelper.HelmChartName,
-		"https://storage.googleapis.com/solo-public-helm", "--force-update")
-	args = append(args, "gloo/gloo",
-		"--version", fromRelease)
-
-	args = append(args, "-n", testHelper.InstallNamespace,
-		"--create-namespace",
-		"--values", valueOverrideFile)
-	if strictValidation {
-		args = append(args, strictValidationArgs...)
-	}
-
-	fmt.Printf("running helm with args: %v\n", args)
-	runAndCleanCommand("helm", args...)
-
-	// Check that everything is OK
-	checkGlooHealthy(testHelper)
-}
-
-// CRDs are applied to a cluster when performing a `helm install` operation
-// However, `helm upgrade` intentionally does not apply CRDs (https://helm.sh/docs/topics/charts/#limitations-on-crds)
-// Before performing the upgrade, we must manually apply any CRDs that were introduced since v1.9.0
-func upgradeCrds(testHelper *helper.SoloTestHelper, crdDir string) {
-	// apply crds from the release we're upgrading to
-	fmt.Printf("Upgrade crds: kubectl apply -f %s\n", crdDir)
-	runAndCleanCommand("kubectl", "apply", "-f", crdDir)
-	// allow some time for the new crds to take effect
-	time.Sleep(time.Second * 10)
-}
-
-func upgradeGloo(testHelper *helper.SoloTestHelper, chartUri string, crdDir string, strictValidation bool, additionalArgs []string) {
-	upgradeCrds(testHelper, crdDir)
-
-	valueOverrideFile, cleanupFunc := getHelmUpgradeValuesOverrideFile()
-	defer cleanupFunc()
-
-	var args = []string{"upgrade", testHelper.HelmChartName, chartUri,
-		"-n", testHelper.InstallNamespace,
-		"--values", valueOverrideFile}
-	if strictValidation {
-		args = append(args, strictValidationArgs...)
-	}
-	args = append(args, additionalArgs...)
-
-	fmt.Printf("running helm with args: %v\n", args)
-	runAndCleanCommand("helm", args...)
-
-	//Check that everything is OK
-	checkGlooHealthy(testHelper)
-}
-
-func uninstallGloo(testHelper *helper.SoloTestHelper, ctx context.Context, cancel context.CancelFunc) {
-	Expect(testHelper).ToNot(BeNil())
-	err := testHelper.UninstallGloo()
-	Expect(err).NotTo(HaveOccurred())
-	_, err = kube2e.MustKubeClient().CoreV1().Namespaces().Get(ctx, testHelper.InstallNamespace, metav1.GetOptions{})
-	Expect(apierrors.IsNotFound(err)).To(BeTrue())
-	cancel()
-}
-
-func getHelmUpgradeValuesOverrideFile() (filename string, cleanup func()) {
-	values, err := ioutil.TempFile("", "values-*.yaml")
-	Expect(err).NotTo(HaveOccurred())
-
-	_, err = values.Write([]byte(`
-global:
-  image:
-    pullPolicy: IfNotPresent
-  glooRbac:
-    namespaced: true
-    nameSuffix: e2e-test-rbac-suffix
-settings:
-  singleNamespace: true
-  create: true
-  replaceInvalidRoutes: true
-gateway:
-  persistProxySpec: true
-gatewayProxies:
-  gatewayProxy:
-    healthyPanicThreshold: 0
-    gatewaySettings:
-      # the KEYVALUE action type was first available in v1.11.11 (within the v1.11.x branch); this is a sanity check to
-      # ensure we can upgrade without errors from an older version to a version with these new fields (i.e. we can set
-      # the new fields on the Gateway CR during the helm upgrade, and that it will pass validation)
-      customHttpGateway:
-        options:
-          dlp:
-            dlpRules:
-            - actions:
-              - actionType: KEYVALUE
-                keyValueAction:
-                  keyToMask: test
-                  name: test
-`))
-	Expect(err).NotTo(HaveOccurred())
-
-	err = values.Close()
-	Expect(err).NotTo(HaveOccurred())
-
-	return values.Name(), func() { _ = os.Remove(values.Name()) }
-}
-
-var strictValidationArgs = []string{
-	"--set", "gateway.validation.failurePolicy=Fail",
-	"--set", "gateway.validation.allowWarnings=false",
-	"--set", "gateway.validation.alwaysAcceptResources=false",
-}
-
-func runAndCleanCommand(name string, arg ...string) []byte {
-	cmd := exec.Command(name, arg...)
-	b, err := cmd.Output()
-	// for debugging in Cloud Build
-	if err != nil {
-		if v, ok := err.(*exec.ExitError); ok {
-			fmt.Println("ExitError: ", string(v.Stderr))
-		}
-	}
-	Expect(err).To(BeNil())
-	cmd.Process.Kill()
-	cmd.Process.Release()
-	return b
-}
-
-func checkGlooHealthy(testHelper *helper.SoloTestHelper) {
-	deploymentNames := []string{"gloo", "discovery", "gateway-proxy"}
-	for _, deploymentName := range deploymentNames {
-		runAndCleanCommand("kubectl", "rollout", "status", "deployment", "-n", testHelper.InstallNamespace, deploymentName)
-	}
-	kube2e.GlooctlCheckEventuallyHealthy(2, testHelper, "90s")
-}
