+++ test/kube2e/upgrades/README.md
@@ -1,3 +1,4 @@
 # Upgrade Tests 
 helm - Pinned version upgrade tests. These tests are pinned to specific upgrades due to found issues 
-multiversion - tests that change upgrade version checks based on current version 
+glooupgrade - tests that change upgrade version checks based on current version 
+
