+++ test/kube2e/upgrades/glooupgrade/upgrade_suite_test.go
@@ -1,4 +1,4 @@
-package upgrade_test
+package glooupgrade_test
 
 import (
 	"os"
