+++ test/kube2e/upgrades/shared_upgrade_tests.go
@@ -1,4 +1,4 @@
-package upgrade
+package upgrades
 
 import (
 	"context"
