+++ test/kube2e/upgrades/upgrade_utils.go
@@ -1,4 +1,4 @@
-package upgrade
+package upgrades
 
 import (
 	"context"
